+++ CHANGELOG.md
@@ -3,8 +3,11 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## 7.2.0
+
 * New property `fillAfterStrokeEnabled` for `Konva.Shape`. See API docs for more information.
 * Fix for `Konva.Transformer` when it may fail to draw.
+* Fix rendering of `TextPath` one more time.
 
 ## 7.1.9